refactor: lift sampling into the backend abstraction (review P2)
- TrainingBackend ABC gains sample() and prepare_for_generation()
- NeMo RL batching (NemoRLBatchAccumulator + batched generate) moves from
  SamplingService into backends/nemo_rl/generation.py; accumulators are
  keyed per model and dropped on delete_model
- Miles SGLang HTTP sampling moves into MilesBackend.sample()
- SamplingService is now backend-agnostic: resolves the handle and
  delegates; no backend_type sniffing, no private backend imports
- Behavior notes: sampling-path errors are BackendError (was RuntimeError);
  create_sampling_client now calls prepare_for_generation (fast no-op when
  generation-ready; validates the SGLang router on Miles)
- Contract tests for the new methods
